daemon: Update data model handling for flat designs
Data model parsing of on target packages needed some update to be able
to parse flat shell designs. Also shell.json is not optional anymore.

Also send a response from daemon to the client socket for each of the
commands so that the appropriate messages are seen on ssh clients.

Signed-off-by: pujag <pujag@xilinx.com>