linux: Make "accel_devices" in accel.json dynamic array
The number of IP devices can vary based on the base design, hence
make the array dynamic and the file descriptor for IP can be requested
using getFD() and passing the string for IP required.

Also introduce a new field for shell.json "load_base_design" which by
default is set to 'yes'. Set it to 'no' in shell.json if the user doesn't
intend to load the base design before loading the accelerators to slot.
This is added since for some usecases the PL is programmed during bootup
with base design and acclerators can be directly loaded to slots.

Signed-off-by: Puja Gupta <pujag@xilinx.com>