dfx-mgr: Add firmware detection capability
The newly added xlnx-firmware-detect script provides a mechanism for detecting
and loading default firmware through dfx-mgr. The script is executed as a
pre-step in the dfx-mgr unit file prior to starting the dfx-mgrd. It uses
fru-print to read the board ID from an EEPROM and based on that searches for a
product specific firmware entry in the firmware search path defined in the
dfx-mgr config file. In case the script is able to detect compatible default
firmware, it will update the dfx-mgr config file accordingly, otherwise the
script will fail but the dfx-mgr daemon is still started and no default firmware
will be loaded.

Signed-off-by: Christian Kohn <christian.kohn@xilinx.com>
Signed-off-by: Shubhangi Mahalle <shubhangi.shrikrushna-mahalle@xilinx.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@xilinx.com>
Acked-by: Sergei Korneichuk <sergei.korneichuk@xilinx.com>