dfx-mgr-client: unload SIHA base on request
The dfx-mgr supports the loading of DFX design which consist of a base
(sometimes called "shell") PL design and then accelerators or "RMs" are
loaded dynamically within RPs within the base design. The loading and
unloading of RMs appear to happen correctly but when a new full or base
bitstream is requested to be loaded, the unload command lacks
functionality to unload the base design to allow a different base/shell
design to be loaded. Request is for "unload" command a hook to unload
the base design.

Accept "base" as an argument for dfx-mgr-client -remove option:
"dfx-mgr-client -remove base" to remove SIHA base if possible.

Signed-off-by: Sergei Korneichuk <sergei.korneichuk@amd.com>
Acked-by: Wesley Skeffington <wesley.skeffington@amd.com>