xlnx-firmware-detect: Remove valid ID check
The current implementation checks the board ID read from EEPROM against a list
of valid IDs. This doesn't scale well as everytime a new board is released, the
valid ID list needs to be appended. It is safe to assume that if a firmware
folder exists on the target board that matches the expected name derived from
the board ID, it is valid. Therefore, remove the valid ID check so this script
does not need to be updated for every new platform.

Signed-off-by: Christian Kohn <chris.kohn@amd.com>